Backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.3
secure-smtplib==0.1.1
SQLAlchemy==2.0.39
starkbank-ecdsa==2.2.0
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
